+++ DUBSON_Googs_Delivery_Program/DestinationInterface.cs
@@ -6,7 +6,7 @@ namespace DUBSON_Goods_Delivery_Program
 {
     public interface DestinationInterface
     {
-       public long Distance { get; }
+       public double Distance { get; }
 
        public string Name { get; }
